app/requirements.txt
scrapy==2.11.0
psycopg2-binary==2.9.9
redis==5.0.1
itemadapter==0.8.0
pymongo==4.5.0.